task: change avro to tagparser
