--- lib/src/avro/implementations/avro_js.dart
@@ -1,83 +1,0 @@
-@JS('avsc.Type')
-library avsc;
-
-import 'dart:typed_data';
-
-import 'package:arweave/src/models/tag.dart';
-import 'package:js/js.dart';
-
-external AvroType get avroType;
-
-AvroType _getTypeSchema() => AvroType.forSchema(Schema(
-      type: 'record',
-      name: 'Tag',
-      fields: [
-        SchemaField(name: 'name', type: 'string'),
-        SchemaField(name: 'value', type: 'string'),
-      ],
-    ));
-
-Uint8List serializeData({required List<Tag> tags}) {
-  var obj = {
-    getTypeForSchema(Schema(
-      type: 'record',
-      name: 'Tag',
-      fields: [
-        SchemaField(name: 'name', type: 'string'),
-        SchemaField(name: 'value', type: 'string'),
-      ],
-    ))
-  };
-  print(obj);
-
-  return _getTypeSchema()
-      .toBuffer(tags.map((tag) => BundleTag(name: tag.name, value: tag.value)));
-}
-
-@JS('Type.forSchema')
-external dynamic getTypeForSchema(Schema schema);
-
-@JS('Type')
-abstract class AvroType {
-  external static dynamic forSchema(Schema schema);
-  external dynamic toBuffer(Object obj);
-}
-
-@JS()
-@anonymous
-class Schema {
-  external String type;
-  external String name;
-  external List<SchemaField> fields;
-
-  external factory Schema({
-    String type,
-    String name,
-    List<SchemaField> fields,
-  });
-}
-
-@JS()
-@anonymous
-class SchemaField {
-  external String name;
-  external String type;
-
-  external factory SchemaField({
-    String name,
-    String type,
-  });
-}
-
-@JS()
-@anonymous
-class BundleTag {
-  external String get name;
-  external String get value;
-
-  // Must have an unnamed factory constructor with named arguments.
-  external factory BundleTag({
-    String name,
-    String value,
-  });
-}
