+++ lib/src/utils/bundle_tag_parser.dart
@@ -2,7 +2,7 @@ import 'dart:typed_data';
 
 import 'package:arweave/src/models/tag.dart';
 
-import 'implementations/avro_js.dart'
+import 'implementations/bundle_tag_parser_js.dart'
     if (dart.library.io) 'implementations/pst_stub.dart' as implementation;
 
 Uint8List serializeTags({required List<Tag> tags}) =>
