+++ lib/src/utils/implementations/bundle_tag_parser_js.dart
@@ -1,0 +1,29 @@
+@JS('tagparser')
+library tagparser;
+
+import 'dart:typed_data';
+
+import 'package:arweave/src/models/tag.dart';
+import 'package:js/js.dart';
+
+@JS('serializeData')
+external Uint8List _serializeData({required Iterable<BundleTag> bundleTags});
+
+Uint8List serializeData({required List<Tag> tags}) {
+  return _serializeData(
+      bundleTags:
+          tags.map((tag) => BundleTag(name: tag.name, value: tag.value)));
+}
+
+@JS()
+@anonymous
+class BundleTag {
+  external String get name;
+  external String get value;
+
+  // Must have an unnamed factory constructor with named arguments.
+  external factory BundleTag({
+    String name,
+    String value,
+  });
+}
